Self Check
Do the elements that contain the text "This should be styled" have the correct styles applied?
Yes, using the chaining combination of class name and type of element all text "This should be styled" were correctly styled.

Do the elements that contain the text "This should be unstyled" have no styles applied?
Yes.